+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## Unreleased
+
+IMPROVEMENTS:
+* resource/list_chart: Added `hideMissingValues` argument to show or hide missing values in the chart.
+
 ## 6.0.0 (October 23, 2020)
 
 IMPROVEMENTS: